requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
apptools==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
configobj==5.0.8
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
discord-py-slash-command==4.2.1
dnspython==2.3.0
email-validator==1.3.1
entrypoints==0.4
envisage==7.0.3
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-Redmail==0.3.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
fonttools==4.38.0
frozenlist==1.3.3
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mayavi==4.8.1
mistune==2.0.4
MouseInfo==0.1.3
multidict==6.0.3
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclassic==0.4.7
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
opencv-python==4.8.0.76
openpyxl==3.0.10
packaging==21.3
Panda3D==1.10.13
panda3d-gltf==0.14
panda3d-simplepbr==0.10
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
Pint==0.20.1
plyer==2.1.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.3
pure-eval==0.2.2
py-cord==2.4.1
PyAutoGUI==0.9.53
pycparser==2.21
pyface==8.0.0
pygame==2.5.1
PyGetWindow==0.0.9
Pygments==2.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
pyserial==3.5
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
PyVISA==1.12.0
pywin32==304
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
redmail==0.5.0
requests==2.28.2
sciparse==0.1.24
scippy==0.1.31
scipy==1.9.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.0
statsmodels==0.13.5
stripe==5.2.0
terminado==0.17.0
tinycss2==1.2.1
tkcalendar==1.6.1
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
traits==6.4.1
traitsui==7.4.3
typing_extensions==4.4.0
urllib3==1.26.14
ursina==5.2.0
vtk==9.2.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
win10toast==0.9
windows-curses==2.3.1
WTForms==3.0.1
yarl==1.8.2
zipp==3.11.0
